Task: 1-20 .travis.yml added
travis file has been added to ensure automatic building of the code on the Travis CI server. A .rosinstall has also been included which states all the dependencies that travis need to install to build the code successfully. The travis file also includes the command to check the code coverage on coveralls.