Task: 3-6 Update Readme
The readme has been updated. Some of the already existing information has been augmented with links. But the instructions to record and play the bag file, and to run the tests have been added. Also, the instructions to call the services have been added. Instructions have been added to use some of the other arguments of the launch file. Note that, the section about the contributors has also been updated.