Task: 3-9 Tests for class Detection changed
According to the change made to class Detection method, the tests have been modified. Note that, for now, the tests are not checking anything. This is because we still could not figure out the error in the code. Once that is resolved, the tests would be properly developed. In other terms, the tests are currently dummy tests.