Task: 3-6 README updated
The structuring of the README file has been updated to better follow the instructions. Also, some minor changes have been made after running the code on a new system. The major change in the README is the addition of instructions to generate the code coverage report. Note that, spell check has also been run on the README file.